reorganized adminManagement files, and started working on design of addEmployee form
