+++ src/App.js
@@ -7,7 +7,7 @@ import Login from "./components/Login";
 import Logout from './components/Logout';
 import UpdateAccount from './components/UpdateAccount';
 import AdminHome from './components/Admin/AdminHome';
-import AdminManagement from './components/Admin/AdminManagement';
+import AdminManagement from './components/Admin/AdminManagementComponents/AdminManagement';
 import UnknownPage from './components/UnknownPage';
 import PrivateRouterUser from './components/PrivateRouterUser'
