+++ src/components/Admin/AdminManagementComponents/EmployeeTable.css
@@ -1,4 +1,4 @@
-
+/* This css sheet is for both the employeetable.js and adminManagement.js files */
 .EmployeeTable {
     border-collapse: collapse;
     margin:0 auto;
@@ -34,26 +34,16 @@
     background-color: #ddd;
   }
   
-  .EmployeeTable button,
-  .AddEmployee button {
-    background-color: #a61a11;
-    color: white;
-    border: none;
-    /* padding: 8px 16px; */
-    cursor: pointer;
-    font-size: 1.2rem;
-    padding: 10px 20px;
-  }
   
-  .EmployeeTable button:hover,
-  .AddEmployee button:hover  {
-    background-color: #c71f16;
-  }
 
-  .AddEmployee {
-    margin-top: 20px;
-    text-align: center;
-  }
+ 
+
+
+
+  
+
+  
+
